requirements.txt
Django==3.1.5
Pillow==8.1.0
boto3==1.16.59
dj-database-url==0.5.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.11.1
django_extensions==3.1.0
django_redis==4.12.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
redis==3.5.3
requests==2.25.1
sentry-sdk==0.19.5
wagtail==2.11.3
wagtailmenus==3.0.2
whitenoise==5.2.0

-e git+https://github.com/jhnbkr/wagtail-modeltranslation@9b3120fd366b8a16e8099f1ccd1c531f8e86e703#egg=wagtail-modeltranslation==0.10.17